Add an example of using nock
