+++ test/test.js
@@ -1,39 +1,23 @@
-var colorize = require('../lib/ColorizeShift');
-var dateFormat = 'ddd, DD MMM YYYY HH:mm:ss ZZ';
-var moment = require('moment');
+var nock = require('nock');
+var WhenIWork = require('wheniwork-unofficial');
+var config = require('../config');
 var assert = require('assert');
 
-var tests = [
-  {
-    time: moment(1456995600000).format(dateFormat),
-    humanReadable: 'make 4am red',
-    color: 'D61F27'
-  }, {
-    time: moment(1457017200000).format(dateFormat),
-    humanReadable: 'make 10am gray',
-    color: 'AAAAAA'
+var base = nock('https://api.wheniwork.com/2');
+var loginResponse = {
+  login: {
+    token: 'FAKETOKEN'
   }
-];
+};
 
-describe('ColorizeShift', function () {
-  describe('colorize(obj, timestamp)', function () {
-    for (var i in tests) {
-      it('should ' + tests[i].humanReadable, function () {
-        var shift = {};
-        shift = colorize(shift, tests[i].time);
-        assert.equal(shift.color, tests[i].color);
-      });
-    }
-  });
+base.post('/login').reply(200, loginResponse);
 
-  describe('colorize(obj)', function () {
-    for (var i in tests) {
-      it('should ' + tests[i].humanReadable, function () {
-        var shift = {start_time: tests[i].time};
-        shift = colorize(shift);
-        assert.equal(shift.color, tests[i].color);
-      });
-    }
+it('should log in and get shifts', function (done) {
+  var api = new WhenIWork(config.wheniwork.api_key, config.wheniwork.username, config.wheniwork.password);
+  base.get('/shifts?').reply(200, {ninja: 'turtle'});
+
+  api.get('shifts', function (res) {
+    assert.equal(res.ninja, 'turtle');
+    done();
   });
 });
-
